# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ultimate Life OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stats/weekly:
    get:
      operationId: getWeeklyStats
      tags: [stats]
      summary: Get weekly statistics
      parameters:
        - in: query
          name: week
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Weekly stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyStats"
  /stats/monthly:
    get:
      operationId: getMonthlyStats
      tags: [stats]
      summary: Get monthly statistics
      parameters:
        - in: query
          name: month
          schema:
            type: string
      responses:
        "200":
          description: Monthly stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyStats"
  /stats/heatmap:
    get:
      operationId: getHeatmap
      tags: [stats]
      summary: Get activity heatmap data
      parameters:
        - in: query
          name: days
          schema:
            type: integer
            default: 365
      responses:
        "200":
          description: Heatmap data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HeatmapDay"

  # ─── POMODORO ────────────────────────────────────────────────────────────────
  /pomodoro/sessions:
    get:
      operationId: getPomodoroSessions
      tags: [pomodoro]
      summary: List pomodoro sessions
      parameters:
        - in: query
          name: date
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Pomodoro sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PomodoroSession"
    post:
      operationId: createPomodoroSession
      tags: [pomodoro]
      summary: Start a pomodoro session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PomodoroSessionInput"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PomodoroSession"
  /pomodoro/sessions/{id}:
    patch:
      operationId: updatePomodoroSession
      tags: [pomodoro]
      summary: Complete or cancel a pomodoro session
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PomodoroSessionUpdate"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PomodoroSession"

  # ─── REPORTS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    DayProgress:
      type: object
      required: [date, completed, total, percent]
      properties:
        date:
          type: string
          format: date
        completed:
          type: integer
        total:
          type: integer
        percent:
          type: number
        grade:
          type: string
    WeeklyStats:
      type: object
      required: [weekStart, weekEnd, days, summary]
      properties:
        weekStart:
          type: string
          format: date
        weekEnd:
          type: string
          format: date
        days:
          type: array
          items:
            $ref: "#/components/schemas/DayProgress"
        summary:
          $ref: "#/components/schemas/WeekSummary"
    WeekSummary:
      type: object
      required: [totalCompleted, totalMissed, avgCompletion, bestDay, worstDay, bestCategory, worstCategory]
      properties:
        totalCompleted:
          type: integer
        totalMissed:
          type: integer
        avgCompletion:
          type: number
        bestDay:
          type: ["string", "null"]
        worstDay:
          type: ["string", "null"]
        bestCategory:
          type: ["string", "null"]
        worstCategory:
          type: ["string", "null"]
        totalPomodoroMinutes:
          type: integer
        streak:
          type: integer
    MonthlyStats:
      type: object
      required: [month, year, weeks, summary]
      properties:
        month:
          type: integer
        year:
          type: integer
        weeks:
          type: array
          items:
            $ref: "#/components/schemas/WeekSummary"
        summary:
          $ref: "#/components/schemas/MonthSummary"
    MonthSummary:
      type: object
      required: [totalCompleted, totalMissed, avgCompletion, topCategory]
      properties:
        totalCompleted:
          type: integer
        totalMissed:
          type: integer
        avgCompletion:
          type: number
        topCategory:
          type: ["string", "null"]
        bottomCategory:
          type: ["string", "null"]
        discoveredHabits:
          type: array
          items:
            type: string
        lostHabits:
          type: array
          items:
            type: string
    HeatmapDay:
      type: object
      required: [date, count, level]
      properties:
        date:
          type: string
          format: date
        count:
          type: integer
        level:
          type: integer

    # ─── POMODORO SCHEMAS ───────────────────────────────────────────────────────
    PomodoroSession:
      type: object
      required: [id, type, durationMinutes, status, startedAt]
      properties:
        id:
          type: integer
        taskId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [focus, short_break, long_break]
        durationMinutes:
          type: integer
        status:
          type: string
          enum: [active, completed, cancelled]
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
    PomodoroSessionInput:
      type: object
      required: [type, durationMinutes]
      properties:
        taskId:
          type: integer
        type:
          type: string
          enum: [focus, short_break, long_break]
        durationMinutes:
          type: integer
    PomodoroSessionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [completed, cancelled]
        completedAt:
          type: string
          format: date-time
        notes:
          type: string

    # ─── REPORT SCHEMAS ─────────────────────────────────────────────────────────
...