# lib/api-zod/src/generated/types/monthlyStats.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Ultimate Life OS API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthSummary } from './monthSummary';
import type { WeekSummary } from './weekSummary';

export interface MonthlyStats {
  month: number;
  year: number;
  weeks: WeekSummary[];
  summary: MonthSummary;
}