# lib/api-zod/src/generated/types/weekSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Ultimate Life OS API
 * OpenAPI spec version: 0.1.0
 */

export interface WeekSummary {
  totalCompleted: number;
  totalMissed: number;
  avgCompletion: number;
  /** @nullable */
  bestDay: string | null;
  /** @nullable */
  worstDay: string | null;
  /** @nullable */
  bestCategory: string | null;
  /** @nullable */
  worstCategory: string | null;
  totalPomodoroMinutes?: number;
  streak?: number;
}